No printer bug fix (#127)
* No printer bug

I moved the code that was causing the crash on first install, or if somehow all printers were removed. I changed it so now if the printer status changes, or on every run of Cura, it checks for a printer rather than just attempting to connect to it.

* Renamed hasPrinter()

hasPrinter() has been renamed to printerCheck for more clarity.