R20.1 new pywim api updates (#150)
* API update for Thor changes

The REST API has been changed and the plugin is now using the correct calls to the API. It is also now building a config file, as well as a login token file so we can store this information on the user's drive, rather than in the preferences within Cura, giving a little more flexibility and ease of testing/adjustment.

* Missed an unused import

Removed QStandardPaths, was one of the ways to path that ended up being more work than needed.

* Requested Changes

Adjusted some variable names, changed back to the preferences system rather than a config file, renamed the token file, adjusted the try/except blocks and fixed a few commas with no spaces.

* Changes for PR

fileExists has been removed, variables correctly renamed, private methods indicated and logic for token file opening improved.

* Fixes to token returning

I had an issue with the writing permissions in creating the token file, as well as the JSON dump() method call. Fixed a situation where the token would have been returned as None.

* gitignore add and PR Changes

I added the token file to git ignore so it shouldn't sync anyone's ticket to the github. We are also handling Thor API errors if an HTTP issue arises.