Subscription System (#175)
* Subscription System

*Now checks for subscription statuses and counts for optimizations/validations.

*Displays popups to user to indicate expired subscription, not enough optimizations/validations

-Still needs popup for expired trial.

* Updates to handle the new status

Subscription status was added, so the code now handles those situations and can properly determine if the account was on a trial or not.

* Fixes

I was not handling the status from the API correctly, that is fixed now.

Logic for handling expired trial is now set and ready to go for testing after 7/28/20 once the account expires.

* Updates to CloudConn

Moved the UI management and messages out of the API class and into their own methods in the CloudConnector class. Now the API only returns information about the subscription.

* PR Changes

Using a proper Enum now.

* Font Size

Ok, after much trial and error I got the text to appear as an HTML heading 2 and resize to larger.

* Changed to Heading 3

HTML tag for the text is now Heading 3, making it smaller.