R20.1 handle getaddrinfo failed (#228)
* Fix for #212

This should catch all the rate limit errors

* Added a timeout system to the API calls

New timeout system gives 3 chances at 20 seconds (currently) for any API call to complete successfully.

Additionally I fixed a weird bug with the subscription system.

* PR Changes

The code to call the API is now moved to its own function that uses lambdas to cut down on code duplication. Thanks to Brady for the assist. Looking much cleaner. Also created an Enum for the connection errors for maintainablility.

* Version Change

* PR Changes

Reformatted the long lines and improved the timeout system so it now has far shorter timeouts but will attempt more times. Also now logging the error that comes out of the Try/Except in the API calls.

* Fixed class variables.

Cleared up my confusion and fixed the change that was requested.

* PR Requests

Cleaned up the messages a bit. Removed the "Failed to cancel job message", decided just indicating internet issues was enough. Added a new message for when the system is attempting to reconnect. Got the messages to clear correctly when the connection comes back, or when a new API function is attempted.

* Version Updates

* PR Request

Added a small internet check to prevent Cura locking. Removed connection interruption messages.

* Ver. update