R20.1 load handler (#206)
* rotate handle for the select tool

* position and scale of handle set

* working slider

* moved the rotation handle into the scene, got the position and orientation working

* rearchited the load handler so it works when adding new loads

* load arrow now part of the rotation tool

* created a utility routine to do the angle between Cura vectors better than they do

* switched the ordering of the tools so we can turn on and off the rotator

* remove rotation tool for normal loads

* changed default arrow direction to towards surface

* fixed a display issue

* code cleanup

* rearchitecture plus flipping direction for normal loads

* turn off rotation handler if needed

* slider with load helper image

* qml cleanup

* versioning

* fixed a bug in project saving/opening for the rotation of loads and created some routines to open a saved result in the new load systems with rotators

* PR notes

* fixes from pr notes

* cleanup

* can rotate the arrow

* flip arrow now works as desired

* hacked in face selection when the rotator should also be active

* minor updates

* cleanup

* some initial property tracking

* Good Start

This needs a little tweaking still, I feel like the way the margins are on the labels for the icons is not going to scale correctly, but as it stands, it's functional on a normal resolution. Had to rework a fair bit of stuff, needed rearranged fairly seriously into more columns and rows...

-Still need to hook the buttons up.

* property trcking is good now

* changed the default surface type

* bug fix

* doing some more hacking to get face selection working

* cleanup

* fixed a bug in open project for cylindrical surfaces

* changed to flat as default

* version update

* pywim implementation

* changed back to flat as default

* Commiting To Merge r20.1

Buttons are coming along. A few more issues to fix.

* Proper merge from r20.1

Things got scrambled in the auto merge. I have the changes fixed now.

* Updates to the signals

Improved from this morning, crashes are gone, other than the Cura crash.

Still needs work on some of the UI updating signals, buttons aren't correctly updating between adding new loads and swapping between existing loads. A fair amount of bugs related to adding loads, switching between loads and selections/directions. The indicator arrow and rotator like to disappear too.

* fixed a bug, and some PR changes

* changed back to flat as default

* no need to re-render on a mouse move event

* Styling and adjustments

I created a function that handles the icons and their enabled status, which has fixed the problems with how it was handling loads from the list and adding a new load. Things now update properly.

I also reworked the QML a little to remove a warning it was spamming. The icons and things are now contained in a ColumnLayout and correctly respaced I believe.

* Version

* PR Changes

A little cleanup, a better use of the enum objects

* Fixed the icon updating

When you click on anchor and back on load or off the model and back on, the icons now update correctly.

* Some PR Notes

* PR Changes - Moved buttons to UM.SimpleButtons

Got the icons changed and some of the duplicated code cut down.

* derived a new property for loadfaces and moved the signal to the mouse unclick event

* PR Changes

* Fixed the bug where the face didn't deselect.

* fix for the weird surface bug on mouse movement

* minor mods

* fixed some bugs, took the padding to 0 on the selection mesh

* moved the renderer into a function

* updated version

* fixed a lot of the rendering bugs

* hopefully fixed the nan arrow direction

* PR notes

* removed meshes from the Preview stage

* PR bug fixes

* fixed the flipping stage issue

* bug fixes from the property handler

* removed commented line

* merge r20.1 and incremented version

* merged r20.1 and incremented version

* fixed a few bugs in switching stages and tools - consolidated a lot of code

* added the selected face to the property tracker

* fixed the problem of not being able to change surfaces after project open

* fixed a bug when switching between load/anchor and not having triangles on the load

* QML is finished for the new buttons

Still have some bugs to iron out, the face selection is not fully hooked up for anchors yet.

* Fixed the crash

This is almost there, one minor bug left that I know of to fix

* Fixed the properties bug

* Ver Change

* PR requests

* PR Request

Fixed else/if formatting issue. Fixed explicit boolean checking.

* fixed issues I found during testing Jake's plugin

* PR requests

* version increment

Co-authored-by: colman <colman.hartinger@tetonsim.com>
Co-authored-by: jake-kirkland-teton <62028020+jake-kirkland-teton@users.noreply.github.com>