R20.2 custom messaging (#353)
* First steps in the new Messaging system

* Removed hide

* SmartSlice Messaging system

This was a learning process. I left in the test messages for right now, for the PR purposes. That should be the last removal after all the requested changes are made and it is ready to roll.

* Cleaned some defs that are unneeded.

* PR Requests

Not fully done yet, but most of the changes are made now.

* Quick Fix

Change prevents a UM.Message object from attempting to show using our system, which causes a crash.

* PR requests

Made some formatting changes, removed an unneeded class. Reworked a few functions and removed addAction override

* PR Button Requests

Changed the buttons that are displayed on our popups to be Cura buttons so they follow the same styling as the Validate/Opt and Cancel/Other buttons.

* ActionButtonAlignment derivation

* Final Test Message Removal