R20.2 problem meshes (#364)
* Messages added for the new results buttons

Added new messages to convey results information to the user.
Adjusted the old "Unable to optimize" message to work better with the new messages.

* Fixed button coloring issue

* Bug fix and PR

I managed to fix a bug with the way the buttons were behaving on a double click.
PR requests to set the text as variables, change the text, and remove the old default cannot optimize popup.

* PR fixes

Spelling error fixed and wording.

* Moved Mesh Generation

* Addition of problem meshes.

A system to display meshes, a system to remove and then redisplay meshes and attaching them to the results meshes buttons.

* Version and Server

Currently using test for these changes. Revert to API once changes are live.

* PR changes or UX and some code changes

* Class move

The meshes have been moved to their own class now within SmartSliceScene.

Unittests are nearly functional, they need a few new mock objects to work right.
The Save/Load system needs a little adjustment.

* Final changes to save/load

Additionally:
Added unittests for new meshes
Fixed functionality for new results buttons
Using show/hide where we can for meshes now

* Fix a test for API

Some tests it changes the newline breaks to <br> some it doesn't.

* Delete SmartSliceDecorator.py

* PR requests and bugfix

Fixed a bug with older validations/optimizes. I thought I was catching them all, but I was only catching if the server had not been updated and the results were not present.

Also moved the SmartSliceMessages into reusable components

Ver update.

* PR requests

Attaching a different build plate to the meshes causes them to not be saved, so it solves the need for removing and replacing the meshes on save.

* Most of the PR changes.

Still need to fix the parent issue and unittests

* Removed Node Constructor.

After testing I can find no reason to create a hanging random node to be the init parent.

* Unittest fix

Changed an escaped newline to a <br>

* Updated Unittests for problem meshes

There is now reasonable coverage for the code that was added for the new problem meshes.

Most PR requests

* Required PR changes to tests

MockModel has moved to the _SmartSliceTestCase class as a method

Meshes can now be generated from a method rather than in each test manually

This includes the double check in Proxy change.

* PR changes

Requested move of modifier mesh property updating into it's own generalized function in utils.

* Static Function Fix and Type Annotations

* Function renaming.

* Missed a file save.

* Removed string logic control

* Created a clear results method

Built a way to fully clear results on file load/stage reset

Cleaned up some situations involving old saves